requirements.txt
flickrapi
pandas
python-dotenv